+++ src/fractal_client/config.py
@@ -14,6 +14,7 @@ Zurich.
 
 from os import getenv
 from pathlib import Path
+from typing import Self
 
 from dotenv import load_dotenv
 
@@ -26,10 +27,15 @@ class Settings:
         self.FRACTAL_PASSWORD: str | None = getenv("FRACTAL_PASSWORD")
         self.FRACTAL_TOKEN_PATH: str | None = getenv("FRACTAL_TOKEN")
 
-        self.FRACTAL_SERVER: str = getenv("FRACTAL_SERVER")
+        self.FRACTAL_SERVER: str | None = getenv("FRACTAL_SERVER")
+        self.FRACTAL_WEB: str | None = getenv("FRACTAL_WEB")
         self.FRACTAL_CACHE_PATH: str = getenv(
             "FRACTAL_CACHE_PATH", str(Path.home() / ".cache/fractal")
         )
+
+    @property
+    def default_token_path(self: Self) -> str:
+        return str(Path(self.FRACTAL_CACHE_PATH, "token.txt"))
 
 
 settings = Settings()