+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 # 2.24.0 (unreleased)
 
 * Commands:
-    * Review authentication flow, including interactively prompting for a token (\#924).
+    * Review authentication flow, including interactively prompting for a token (\#924, \#929).
 * Internal:
     * Refactor CLI parser (\#923).
 * Documentation: